+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased]
 
 - feat: Add --message-as-blocks option
-- fix: require libraries
+- fix: Require libraries
+- style: Enable RuboCop new cops and fix offenses
 
 ## [v2.0.1] - 2024-01-17